+++ __tests__/syserrors-tests.js
@@ -6,103 +6,103 @@ jest.unmock( '../dist/class/SolidObject' )
 jest.unmock( '../dist/lib/SolidText' )
 jest.unmock( '../dist/lib/SysErrors' )
 
-const path = require( 'path' )
-let SysErrors
+var path = require( 'path' )
+var SysErrors = void 0
 
-describe( 'lib/SysErrors', () => {
-  describe( 'constructor', () => {
-    beforeEach( () => {
+describe( 'lib/SysErrors', function () {
+  describe( 'constructor', function () {
+    beforeEach( function () {
       SysErrors = require( '../dist/lib/SysErrors')
     })
 
-    it( 'should initialize SysErrors', () => {
-      const testProps = {
+    it( 'should initialize SysErrors', function () {
+      var testProps = {
         lang: 'it',
         includes: [ '/path/to/custom/errdef' ]
       }
 
       SysErrors.setOptions( testProps )
 
-      const options = SysErrors.__get__('options')
+      var options = SysErrors.getOptions()
 
       expect( testProps.lang ).toEqual( options.lang )
       expect( testProps.includes ).toEqual( options.includes )
     })
 
-    it( 'should change properties after initialization', () => {
-      const testProps = {
+    it( 'should change properties after initialization', function () {
+      var testProps = {
         lang: 'pl',
         includes: [ '/path/to/custom/SysErrors' ]
       }
       SysErrors.setOptions( testProps )
-      const options = SysErrors.__get__('options')
+      var options = SysErrors.getOptions()
       expect( testProps.lang ).toEqual( options.lang )
       expect( testProps.includes ).toEqual( options.includes )
     })
 
-    it( 'should convert single include string to string[]', () => {
-      const testProps = {
+    it( 'should convert single include string to string[]', function () {
+      var testProps = {
         includes: '/path/to/custom/SysErrors'
       }
 
       SysErrors.setOptions( testProps )
-      const options = SysErrors.__get__('options')
-      const expected = [ '/path/to/custom/SysErrors' ]
-      const actual = options.includes
+      var options = SysErrors.getOptions()
+      var expected = [ '/path/to/custom/SysErrors' ]
+      var actual = options.includes
 
       expect( expected ).toEqual( actual )
     })
   })
 
-  describe( 'definitionsDir', () => {
-    beforeEach( () => {
+  describe( 'definitionsDir', function () {
+    beforeEach( function () {
       SysErrors = require( '../dist/lib/SysErrors')
     })
 
-    it( 'should get error definitions dir', () => {
-      const LANG = 'it'
+    it( 'should get error definitions dir', function () {
+      var LANG = 'it'
       SysErrors.setOptions({ lang: LANG })
-      const expected = path.join( __dirname, '../dist/definitions/' + LANG )
-      const actual = SysErrors.__get__( 'definitionsDir' )()
+      var expected = path.join( __dirname, '../dist/definitions/' + LANG )
+      var actual = SysErrors.definitionsDir()
       expect( expected ).toEqual( actual )
     })
   })
 
-  describe( 'includeDirs', () => {
-    beforeEach( () => {
+  describe( 'includeDirs', function () {
+    beforeEach( function () {
       SysErrors = require( '../dist/lib/SysErrors')
     })
 
-    it( 'should get error included dir', () => {
-      const LANG = 'it'
-      const INCLUDES = [ '/path/to/include' ]
+    it( 'should get error included dir', function () {
+      var LANG = 'it'
+      var INCLUDES = [ '/path/to/include' ]
       SysErrors.setOptions({ lang: LANG, includes: INCLUDES })
 
-      const expected = [ path.join( '/path/to/include', LANG ) ]
-      const actual = SysErrors.__get__('includeDirs')()
+      var expected = [ path.join( '/path/to/include', LANG ) ]
+      var actual = SysErrors.includeDirs()
 
       expect( expected ).toEqual( actual )
     })
   })
 
-  describe( 'createPropsFrom', () => {
-    beforeEach( () => {
+  describe( 'createPropsFrom', function () {
+    beforeEach( function () {
       SysErrors = require( '../dist/lib/SysErrors')
     })
 
-    it( 'should have error defined as inner props', () => {
-      const testError = new Error( 'test error' )
-      const props = SysErrors.createPropsFrom( testError )
+    it( 'should have error defined as inner props', function () {
+      var testError = new Error( 'test error' )
+      var props = SysErrors.createPropsFrom( testError )
 
       expect( props.hasOwnProperty('inner') ).toBeTruthy()
       expect( props.inner ).toBeDefined()
       expect( props.inner ).toEqual( testError )
     })
 
-    it( 'should return a Solid Error', () => {
-      const testError = new Error()
+    it( 'should return a Solid Error', function () {
+      var testError = new Error()
 
-      const expected = {
+      var expected = {
         code: 'EERR',
         errno: -100,
         name: 'Error',
@@ -110,22 +110,26 @@ describe( 'lib/SysErrors', () => {
         message: 'An Error occurred.',
         inner: testError
       }
-      const expectedKeys = Object.keys( expected ).sort()
+      var expectedKeys = Object.keys( expected ).sort()
 
-      const actual = SysErrors.createPropsFrom( testError )
-      const actualKeys = Object.keys( actual ).sort()
+      var actual = SysErrors.createPropsFrom( testError )
+      var actualKeys = Object.keys( actual ).sort()
 
-      const expectedValues = expectedKeys.map( key => expected[ key ] )
-      const actualValues = actualKeys.map( key => actual[ key ] )
+      var expectedValues = expectedKeys.map( function (key) {
+        return expected[ key ]
+      })
+      var actualValues = actualKeys.map( function (key) {
+        return actual[ key ]
+      })
 
       expect( expectedKeys ).toEqual( actualKeys )
       expect( expectedValues ).toEqual( actualValues )
     })
 
-    it( 'should return a Solid Error with custom message', () => {
-      const TEXT = 'Custom test message'
-      const testError = new Error( TEXT )
-      const expected = {
+    it( 'should return a Solid Error with custom message', function () {
+      var TEXT = 'Custom test message'
+      var testError = new Error( TEXT )
+      var expected = {
         code: 'EERR',
         errno: -100,
         name: 'Error',
@@ -133,23 +137,27 @@ describe( 'lib/SysErrors', () => {
         message: TEXT,
         inner: testError
       }
-      const expectedKeys = Object.keys( expected ).sort()
+      var expectedKeys = Object.keys( expected ).sort()
 
-      const actual = SysErrors.__get__('createPropsFrom')( testError )
-      const actualKeys = Object.keys( actual ).sort()
+      var actual = SysErrors.createPropsFrom( testError )
+      var actualKeys = Object.keys( actual ).sort()
 
-      const expectedValues = expectedKeys.map( key => expected[ key ] )
-      const actualValues = actualKeys.map( key => actual[ key ] )
+      var expectedValues = expectedKeys.map( function (key) {
+        return expected[ key ]
+      })
+      var actualValues = actualKeys.map( function (key) {
+        return actual[ key ]
+      })
 
       expect( expectedKeys ).toEqual( actualKeys )
       expect( expectedValues ).toEqual( actualValues )
     })
 
-    it( 'should return a Solid Error with custom name', () => {
-      const ERROR_NAME = 'TestError'
-      const testError = new Error()
+    it( 'should return a Solid Error with custom name', function () {
+      var ERROR_NAME = 'TestError'
+      var testError = new Error()
       testError.name = ERROR_NAME
-      const expected = {
+      var expected = {
         code: 'EERR',
         errno: -100,
         name: ERROR_NAME,
@@ -157,14 +165,14 @@ describe( 'lib/SysErrors', () => {
         readableName: 'Test Error',
         inner: testError
       }
-      const actual = SysErrors.createPropsFrom( testError )
+      var actual = SysErrors.createPropsFrom( testError )
 
       expect( expected ).toEqual( actual )
     })
 
-    it( 'should give precedence to user defined error', () => {
-      const testError = new RangeError()
-      const expected = {
+    it( 'should give precedence to user defined error', function () {
+      var testError = new RangeError()
+      var expected = {
         code: 'TRNG',
         errno: -201,
         name: 'RangeError',
@@ -173,19 +181,23 @@ describe( 'lib/SysErrors', () => {
         explain: 'This message should show up only during tests.',
         inner: testError
       }
-      const expectedKeys = Object.keys( expected ).sort()
+      var expectedKeys = Object.keys( expected ).sort()
 
       SysErrors.setOptions({
         lang: 'en',
         includes: path.join( __dirname, './support/errorDefinitions' )
       })
 
-      const actual = SysErrors.createPropsFrom( testError )
-      const actualKeys = Object.keys( actual ).sort()
+      var actual = SysErrors.createPropsFrom( testError )
+      var actualKeys = Object.keys( actual ).sort()
 
-      const expectedValues = expectedKeys.map( key => expected[ key ] )
+      var expectedValues = expectedKeys.map( function (key) {
+        return expected[ key ]
+      })
 
-      const actualValues = actualKeys.map( key => actual[ key ] )
+      var actualValues = actualKeys.map( function (key) {
+        return actual[ key ]
+      })
 
       expect( expectedKeys ).toEqual( actualKeys )
       expect( expectedValues ).toEqual( actualValues )