+++ src/class/SolidObject.js
@@ -11,11 +11,49 @@
  * @module class/SolidObject
  * @author    Roberto Mauro <erremauro@icloud.com>
  * @license   MIT License. See LICENSE file in the root directory.
- * @version   0.1.0
+ * @version   0.1.4
  * @since     0.3.1
  *
  * @flow
  */
+
+/**
+ * Generate a standard guid
+ * @memberOf module:class/SolidObject
+ * @function
+ * @inner
+ * @returns {undefined}
+ */
+function guid (): string {
+  const s4 = (): string => Math
+    .floor( ( 1 + Math.random() ) * 0x10000 )
+    .toString(16)
+    .substring(1)
+
+  return s4() + s4() + '-' + s4() + '-' + s4() + '-' +
+  s4() + '-' + s4() + s4() + s4();
+}
+
+/**
+ * SolidObject readonly property symbols
+ * @type {Object}
+ * @memberOf module:class/SolidObject
+ * @inner
+ * @property {Symbol} state
+ * @property {Symbol} props
+ * @property {Symbol} key
+ * @property {Symbol} defaults
+ * @property {Symbol} solid
+ * @since 0.1.4
+ * @version 0.1.0
+ */
+const Symbols = {
+  state: Symbol( '__state__' ),
+  props: Symbol(' __props__' ),
+  key: Symbol( '__key__' ),
+  defaults: Symbol( '__defaults__' ),
+  solid: Symbol( '__solid__' )
+}
 
 /**
  * @class
@@ -68,17 +106,62 @@
  *
  * @param  {?Object} [props] The object properties
  * @param  {?Object} [defaultProps] Object default properties
- * @version   0.1.0
+ * @version   0.1.3
  * @since     0.3.1
  */
 class SolidObject {
-  props: Object
-  defaultProps: Function;
-
   constructor( props?: any, defaultProps?: any ) {
-    this.defaultProps = (): any => defaultProps
-    this.props = defaultProps || {}
+    this[ Symbols.solid ] = true
+    this[ Symbols.key ] = guid()
+    this[ Symbols.props ] = defaultProps || {}
+    this[ Symbols.state ] = {}
+    this[ Symbols.defaults ] = defaultProps
     this.setProps( props )
+  }
+
+  /**
+   * Tells if it's a solid object.
+   * @type {boolean}
+   * @readOnly
+   * @memberOf module:class/SolidObject~SolidObject
+   */
+  get solid(): string {
+    return this[ Symbols.solid ]
+  }
+
+  /**
+   * SolidObject unique key
+   * @type {string}
+   * @readOnly
+   * @memberOf module:class/SolidObject~SolidObject
+   * @since 0.3.1
+   * @version 0.1.3
+   */
+  get key(): string {
+    return this[ Symbols.key ]
+  }
+
+  /**
+   * SolidObject properties
+   * @type {Object}
+   * @readOnly
+   * @memberOf module:class/SolidObject~SolidObject
+   * @since 0.3.1
+   * @version 0.1.3
+   */
+  get props (): Object {
+    return Object.assign( {}, this[ Symbols.props ] )
+  }
+
+  /**
+   * SolidObject internal state object.
+   * @type {Object}
+   * @memberOf module:class/SolidObject~SolidObject
+   * @since 0.3.1
+   * @version 0.1.3
+   */
+  get state(): Object {
+    return Object.assign( {}, this[ Symbols.state ] )
   }
 
   /**
@@ -113,7 +196,7 @@ class SolidObject {
 
     if ( this.propsShouldUpdate() ) {
       props = this.propsWillUpdate( props )
-      this.props = Object.assign( this.props,  props )
+      this[ Symbols.props ] = Object.assign( this.props,  props )
       this.propsDidUpdate()
     }